feyzains/feyzains/feyzainsaat_django/requirements.txt
asgiref==3.8.1
dateutils==0.6.12
Django==5.1.7
django-auditlog==3.0.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1